% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_clean_data.R
\name{hugo_clean_data}
\alias{hugo_clean_data}
\title{This function cleans data.
Funkcja przyjmuje jako argument data.frame z danymi
1. imputuje braki danych
2. w zmiennych skośnych obcina wartości odstające do pewnego kwantyla
3. spłaszcza zmienne factorowe
Funkcja zapisuje do katalogu data zbiór z wyczyszczonymi danymi i informuje o tym użytkownika.}
\usage{
hugo_clean_data(data)
}
\arguments{
\item{data}{Data.}

\item{quantile}{Kwantyl do jakiego obcina skośne zmienne numeryczne}

\item{prop}{Frakcja występowania, poniżej której wartość zmiennej katogorycznej jest uznana za rzadką, i łączona z innymi rzadkimi w jeden factor}
}
\description{
This function cleans data.
Funkcja przyjmuje jako argument data.frame z danymi
1. imputuje braki danych
2. w zmiennych skośnych obcina wartości odstające do pewnego kwantyla
3. spłaszcza zmienne factorowe
Funkcja zapisuje do katalogu data zbiór z wyczyszczonymi danymi i informuje o tym użytkownika.
}
\author{
Eliza Kaczorek
}
