% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_continue_investigation.R
\name{hugo_continue_investigation}
\alias{hugo_continue_investigation}
\title{Continue Hugo investigation}
\usage{
hugo_continue_investigation(path = NULL, session_name = NULL, envir = .GlobalEnv)
}
\arguments{
\item{path}{name of a directory in which the investigation is stored}

\item{session_name}{name of a session to be restored}

\item{envir}{an environment to load variables. Global Environment by default.}
}
\description{
Loads all important variables and all packages from saved investigation.
The investigation can be saved with  \code{hugo_save_investigation} function.
}
\details{
Each investigation can consist of multiple sessions and thus
if \code{session_name} is not provided, the function either loads the data if only one session exists or asks user
to type the session name which is going to be restored. The data is loaded from '\code{resources/{session_name}}' folder, which
contains two files: \code{variables} and \code{packages}. The first one keeps all important variables, while
the second one is used to load all packages which were attached during the saved investigation. The function
tries to load compatible versions of packages, or, if it is not possible, installs latest versions of troublesome packages.
}
\examples{
hugo_start_investigation("iris")
## New investigation started in  iris
my_boxplot <- boxplot(iris$Sepal.Length)
hugo_save_investigation("iris_session1")
## iris_session1  succesfully saved.
remove(my_boxplot)
hugo_continue_investigation("iris")
## Loading required packages to continue investigation ...
## Session iris_session1 from investigation iris sucessfully restored.
}
\seealso{
\code{\link{hugo_save_investigation}}, \code{\link{hugo_start_investigation}}
}
\author{
Mateusz Kobylka
}
