% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_memorise.R
\name{hugo_memorise}
\alias{hugo_memorise}
\title{Saves an object as .rda file}
\usage{
hugo_memorise(obj)
}
\arguments{
\item{obj}{name of an object which will be saved as .rda file}
}
\description{
Saves an object as .rda file
}
\examples{
hugo_memorise()

}
\author{
Anna Gierlak
}
