% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_memorise_plot.R
\name{hugo_memorise_plot}
\alias{hugo_memorise_plot}
\title{Saving plot as .rda, pdf and png}
\usage{
hugo_memorise_plot(plot, name)
}
\arguments{
\item{name}{}
}
\description{
Saving plot as .rda, pdf and png
}
\examples{
hugo_memorise_plot()


}
\author{
Joanna Golawska golawskaj
}
