% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_memorise_table.R
\name{hugo_memorise_table}
\alias{hugo_memorise_table}
\title{Saves an object as: .rda file, .md file and a markdown table as xlsx and docx}
\usage{
hugo_memorise_table(table)
}
\description{
Saves an object as: .rda file, .md file and a markdown table as xlsx and docx
}
\author{
Maria Piliszek
}
