% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_read_data.R
\name{hugo_read_data}
\alias{hugo_read_data}
\title{Reads data to R in many ways}
\usage{
hugo_read_data(path)
}
\arguments{
\item{path}{name of a directory where is file to read}
}
\description{
Reads data to R in many ways
}
\examples{
hugo_read_data()

}
\author{
Dariusz Komosinski
}
