% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_save_investigation.R
\name{hugo_save_investigation}
\alias{hugo_save_investigation}
\title{Save variables and information about packages}
\usage{
hugo_save_investigation(variables = NULL, session_name = NULL)
}
\arguments{
\item{variables}{A character vector containing the names of objects to be saved.
If variables = NULL, all variables in Global Environment are saved.}

\item{session_name}{name of directory in which the session will be saved.
If session_name = NULL, the dafault name is set.}
}
\description{
hugo_save_investigation writes variables and packages used in session.
Written data can be read back at later date by using function hugo_continue_investigation.
}
\examples{
\dontrun{hugo_start_investigation("Example")
model <- lm(Petal.Width~., data = iris)
res <- model$residuals
petal.pred <- predict(model, newdata = iris)
hugo_save_investigation()
# Or select variables to be written out and set name of session's directory:
hugo_save_investigation( variables = c('petal.pred', 'res'), session_name = 'IrisModel')}
}
\author{
Monika Chudek
}
