% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_share_object.R
\name{hugo_share_object}
\alias{hugo_share_object}
\title{Share any object using github}
\usage{
hugo_share_object(object, passphrase = NULL)
}
\arguments{
\item{object}{object to share}

\item{passphrase}{object will be encrypted using provided passphrase}
}
\description{
Share any object using github
}
\examples{
# TODO add some examples
hugo_share_object(iris)
}
\author{
Mateusz Krubiński, Kamil Romaszko
}
