% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_share_secret.R
\name{hugo_share_secret}
\alias{hugo_share_secret}
\title{Encrypt and share any object}
\usage{
hugo_share_secret(object, passphrase)
}
\arguments{
\item{object}{object to share}

\item{passphrase}{object will be encrypted using provided passphrase}
}
\description{
Encrypt and share any object
}
\examples{
hugo_share_secret(iris, "secret")
}
\author{
Kamil Romaszko
}
