% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_show_statistics.R
\name{hugo_show_statistics}
\alias{hugo_show_statistics}
\title{Gives the brief overview of the present hugo statistics.}
\usage{
hugo_show_statistics()
}
\description{
hugo_show_statistics() returns list, which consist of beginning of current research as creationDate,
number of saved objects as objects, number of loaded data sets as dataSets
and number of instructions as instructions.
}
\examples{
hugo_show_statistics()


}
\author{
Joanna Jablonska
}
