% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_start_investigation.R
\name{hugo_start_investigation}
\alias{hugo_start_investigation}
\title{Starts a New Hugo Investigation}
\usage{
hugo_start_investigation(path = NULL)
}
\arguments{
\item{path}{name of a directory in which the investigation will be stored}
}
\description{
Starts a New Hugo Investigation
}
\examples{
hugo_start_investigation()

}
\author{
Przemyslaw Biecek
}
