% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_summarise_data.R
\name{hugo_summarise_data}
\alias{hugo_summarise_data}
\title{Create report with summary of variables in dataset.}
\usage{
hugo_summarise_data(data, overwrite_params = TRUE)
}
\arguments{
\item{data}{R dataset which you wish to summary. This dataset should be of class \code{data.frame},
\code{tibble} or \code{matrix}.}

\item{overwrite_params}{logical. If \code{TRUE} (the default) parameters used to generate report will be saved.}
}
\value{
The function does not return anything. Its side effect (the production
of a data report) is the reason for running the function.
}
\description{
Creates  the report 'html' or 'pdf' in the directory chosen at the beginning of the current
Hugo Investigation in the subfolder 'gallery' in .rda file and a summary in .md file.
This reports summarizes the contents of a dataset
and flags potential problems.
To generate report is used function \code{dataMaid::makeDataReport}.
}
\details{
While using hugo_summarise_data user is prompted and asked to choose additional parameters - report format , decide to overwrite existing report,
decide to automatically open report ,
decide whether numeric and integer variables with
less than 5 unique values are treated as factor variables
 and type max decimals .

If hugo remembers previuos answer, user will be prompted and asked if he wishes to use previous
parameters or choose new ones.
}
\examples{
\dontrun{
# create report for iris dataset
hugo_summarise_data(data = iris)
# create report for cars dataset with no overwrite parameters
hugo_summarise_data(data = cars, overwrite_params = FALSE)
}

}
