% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_train_model.R
\name{hugo_train_model}
\alias{hugo_train_model}
\title{Training binary classification model and choose the best one}
\usage{
hugo_train_model(data, formula)
}
\arguments{
\item{data}{an optional \code{data.frame} containing the variables in the model}

\item{formula}{formula describing the model to be fitted}
}
\value{
Function returns trained models and assessment of how these models are good.
}
\description{
This function for the given formula and data builds logistic regressions models, 
decision trees, gbm model. It saves models to the models catalog and as a result
returns trained models and evaluation of how these models are good.
}
\examples{
\dontrun{
data("PimaIndiansDiabetes")
formula <- diabetes~.
hugo_train_model(PimaIndiansDiabetes, formula)
}

}
\author{
Anna Kozak
}
