% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_clean_data.R
\name{hugo_clean_data}
\alias{hugo_clean_data}
\title{Clean your data}
\usage{
hugo_clean_data(data, prop = 0.05)
}
\arguments{
\item{data}{\code{data.frame} to clean}

\item{prop}{proportion of occurence of the level in a categorical variable which decides which levels are rare}
}
\value{
\code{data.frame} that has been cleaned
}
\description{
This function fills missing values - a median for numeric variables and a mode for categorical variables (factors).
Additionally, the outliers from numeric variables are replaced according to the IQR rule for outliers.
In factors rare levels are merged into 'Other' level.
}
\examples{
\dontrun{
# Dataset in base R: airquality
# There are 44 missing values
sum(is.na(airquality))

hugo_clean_data(airquality)
# The data was cleaned.

# Two original rows from data:

# Ozone Solar.R  Wind Temp Month Day
#     8      19  20.1   61     5   9
#     NA      NA 14.3   56     5   5

# After cleaning:

# Ozone Solar.R  Wind Temp Month Day
#     8      19 17.65   61     5   9
#  31.5     205 14.30   56     5   5

# We can see that the outlier in 'Wind' was
# replaced by the value Q3+1.5*IGR for this column.
# Missing values were replaced with medians.
}

}
\author{
Eliza Kaczorek
}
