% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_memorise.R
\name{hugo_memorise}
\alias{hugo_memorise}
\title{Saving objects as .rda files}
\usage{
hugo_memorise(object = NULL)
}
\arguments{
\item{object}{an object to be saved as .rda file}
}
\description{
The \code{hugo_memorise} function saves an object as .rda file in the 'memory' subdirectory.
If the object of the same name has been already saved, it gives the opportunity either
to overwrite it or to stop the function execution.
}
\examples{
\dontrun{
a <- 1:5
hugo_memorise(a)

b <- list('abcdef', seq(0,5,0.1), c(TRUE,FALSE))
hugo_memorise(b)
}
}
\author{
Anna Gierlak
}
