% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_memorise_model.R
\name{hugo_memorise_model}
\alias{hugo_memorise_model}
\title{Saves a model}
\usage{
hugo_memorise_model(model = NULL, name = substitute(model))
}
\arguments{
\item{model}{a model to be saved}

\item{name}{a name of files}
}
\description{
Saves a model in the directory chosen at the beginning of the current
Hugo Investigation in the subfolder 'gallery' in .rda file and a summary in .md file
}
\examples{
\dontrun{mod.lm <- lm(Sepal.Length ~ Petal.Length, data = iris)
hugo_memorise_model(mod.lm, name = "model.iris")
# Or save as "mod.lm":
hugo_memorise_model(mod.lm)}
}
\author{
Malgosia Lazecka
}
