% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_save_investigation.R
\name{hugo_save_investigation}
\alias{hugo_save_investigation}
\title{Save variables and information about packages}
\usage{
hugo_save_investigation(variables = NULL, session_name = NULL,
  envir = .GlobalEnv)
}
\arguments{
\item{variables}{a character vector containing names of important variables to be saved.
By default \code{variables = NULL}, what means that all the variables in environment \code{envir} are saved.}

\item{session_name}{name of directory in which the session will be saved.
If session_name = NULL, the dafault name is set.}

\item{envir}{an environment to search for objects to be saved. Global Environment by default.}
}
\description{
The \code{hugo_save_investigation} saves variables and versions of packages used in investigation.
Data is stored in subfolder \code{resources/session_name} in two separate files: \code{variables} and \code{packages}.
Restoring data and loading saved versions of packages could be done by using function hugo_continue_investigation.
}
\examples{
\dontrun{hugo_start_investigation("Example")
model <- lm(Petal.Width~., data = iris)
res <- model$residuals
petal.pred <- predict(model, newdata = iris)
hugo_save_investigation()

# or select variables to be written out and set name of session's directory:
hugo_save_investigation( variables = c('petal.pred', 'res'), session_name = 'IrisModel')

# or select the environment that contains variables for saving
e <- new.env()
e$a <- FALSE
e$b <- "a"
expect_output(hugo_save_investigation(envir = e))
}
}
\author{
Monika Chudek
}
