% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_save_investigation.R
\name{hugo_save_investigation}
\alias{hugo_save_investigation}
\title{Saves all important variable and all packages used in project
-- There isnt't info about packages implemented yet.}
\usage{
hugo_save_investigation(session_name = NULL)
}
\description{
Saves all important variable and all packages used in project
-- There isnt't info about packages implemented yet.
}
\examples{
hugo_save_investigation()
}
\author{
Monika Chudek
}
