% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_share_object.R
\name{hugo_share_object}
\alias{hugo_share_object}
\title{Share any object using github}
\usage{
hugo_share_object(object)
}
\arguments{
\item{object}{object to share}
}
\description{
Share any object using github
}
\examples{
# TODO add some examples
hugo_share_object(iris)
}
\author{
MatKru (Mateusz Krubiński)
}
