% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_show_history.R
\name{hugo_show_history}
\alias{hugo_show_history}
\title{Showes all the comends which have been used}
\usage{
hugo_show_history()
}
\description{
Showes all the comends which have been used
}
\examples{
hugo_show_history()
}
\author{
Maciej Kurek
}
