% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_show_history.R
\name{hugo_show_history}
\alias{hugo_show_history}
\title{Dispays history of hugo investigation.}
\usage{
hugo_show_history()
}
\description{
Each usage of any hugo function adds an entry about it to a hidden variable in hugo environment 
which contains the whole history of your investigation.
hugo_show_history(), displays all the entries with additions of paths to files created by functions hugo_memorize*.
}
\author{
Maciej Kurek
}

