% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_summarise_data.R
\name{hugo_summarise_data}
\alias{hugo_summarise_data}
\title{Create report with variables summary in dataset}
\usage{
hugo_summarise_data(data, earlier_settings = TRUE, output = "pdf",
  replace = TRUE, mode = "check", report_title = NULL)
}
\arguments{
\item{data}{The dataset to be summarised. This dataset should be of class \code{data.frame},
\code{tibble} or \code{matrix}. If it is of classs \code{matrix}, it will be converted to a
\code{data.frame}.}

\item{earlier_settings}{A logical. If \code{TRUE} (the default) and this function was used in investigation, saved parameters will be used.
Parameters given by user are overwrite by saved arguments.}

\item{output}{Output format. Options are \code{"pdf"}, \code{"word"} (.docx) and \code{"html"}. This parameters If \code{NULL} (the default),
the output format depends two sequential checks. First, whether a LaTeX installation is available,
in which case \code{pdf} output is chosen. Secondly, if no LaTeX installation
is found, then if the operating system is Windows, \code{word} output is used. Lastly, if neither of these
checks are positive, \code{html} output is used.}

\item{replace}{akjs}

\item{mode}{Vector of tasks to perform among the three categories "summarize", "visualize" and "check".
The default, \code{c("summarize", "visualize", "check")}, implies that all three steps are
performed. The steps selected in \code{mode} will be performed for each variable in
\code{data} and their results are presented in the second part of the outputtet data report.
The "summarize" step is responsible for creating the summary table,
the "visualize" step is responsible for creating the plot and the "check" step is responsible
for performing checks on the variable and printing the results if any problems are found.}

\item{report_title}{A text string. If supplied, this will be the printed title of the
report. If left unspecified, the title with the name of the supplied dataset.}
}
\value{
The function does not return anything. Its side effect (the production
of a data report) is the reason for running the function.
}
\description{
Creates  the report 'html' or 'pdf' in the directory chosen at the beginning of the current
Hugo Investigation in the subfolder 'gallery' in .rda file and a summary in .md file.
This reports summarizes the contents of a dataset
and flags potential problems. The potential problems are identified by
running a set of class-specific validation checks, so that different
checks are performed on different variables types.
}
