% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_summarise_data.R
\name{hugo_summarise_data}
\alias{hugo_summarise_data}
\title{This function creates html or pdf report which summarise variables distribution.
This function uses package dataMaid to create  report.
User can create list of paramters for function makeDataReport()
and in next use this parmeters will be use in default.}
\usage{
hugo_summarise_data(data, ...)
}
\arguments{
\item{...}{}
}
\description{
This function creates html or pdf report which summarise variables distribution.
This function uses package dataMaid to create  report.
User can create list of paramters for function makeDataReport()
and in next use this parmeters will be use in default.
}
