% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_summarise_data.R
\name{hugo_summarise_data}
\alias{hugo_summarise_data}
\title{Create report with summary of variables in dataset.}
\usage{
hugo_summarise_data(data, overwrite_params = TRUE)
}
\arguments{
\item{data}{R dataset which you wish to summary. This dataset should be of class \code{data.frame},
\code{tibble} or \code{matrix}.}

\item{overwrite_params}{logical. If \code{TRUE} (the default) parameters used to generate report will be saved.}
}
\value{
The function does not return anything. Its side effect (the production
of a data report) is the reason for running the function.
}
\description{
Creates  the report 'html' or 'pdf' in the directory chosen at the beginning of the current
Hugo Investigation in the subfolder 'gallery' in .rda file and a summary in .md file.
This reports summarizes the contents of a dataset
and flags potential problems.
To generate report is used function \code{dataMaid::makeDataReport}.
}
\details{
This function is interactive with user. With first usage user has to type parameters, in next usage user
can choose previous settings or type new. User can change filename and title of report.
User can change output format (\code{output}), decide to overwrite output (\code{replace_output}),
  decide whether numeric and integer variables with
less than 5 unique values are treated as factor variables (\code{smart_factor}),
type max decimals (\code{max_decimals}) and decide to automatically open report (\code{open_result}).
}
