% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_train_model.R
\name{hugo_train_model}
\alias{hugo_train_model}
\title{Starts a New Hugo Investigation}
\usage{
hugo_train_model(data, formula)
}
\description{
Starts a New Hugo Investigation
}
\examples{
hugo_train_model()

}
\author{
Anna Kozak
}
