% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_train_model.R
\name{hugo_train_model}
\alias{hugo_train_model}
\title{Training binary classification model and choose the best one}
\usage{
hugo_train_model(data, formula)
}
\arguments{
\item{data}{an optional data frame containing the variables in the model}

\item{formula}{formula describing the model to be fitted}
}
\description{
Training binary classification model and choose the best one
}
\examples{
\dontrun{
data("PimaIndiansDiabetes")
formula <- "diabetes~."
hugo_train_model(PimaIndiansDiabetes, formula)
}

}
\author{
Anna Kozak
}
